Make paint_brush size two dimensional
The size of the paint brush .png image we load in was given as a
one-dimension value which means a single value was used for the size of
the y-dimension and x-dimension. This commit makes sure that we use two
values, one for each dimension, instead which allows us to properly
resize the paint brush as we want.